CI PIPELINE: repetitive files combines
