+++ src/components/ciPipeline/SourceMaterials.tsx
@@ -5,7 +5,6 @@ import { Link } from 'react-router-dom'
 import ReactSelect from 'react-select'
 import error from '../../assets/icons/misc/errorInfo.svg'
 import git from '../../assets/icons/git/git.svg'
-import { reactSelectStyles } from './ciPipeline.util'
 import { getCustomOptionSelectionStyle } from '../v2/common/ReactSelect.utils'
 import { DropdownIndicator } from '../charts/charts.util'
 import { ReactComponent as Info } from '../../assets/icons/ic-info-outline-purple.svg'
@@ -13,6 +12,7 @@ import { ConfigureWebhook } from './ConfigureWebhook'
 import { MaterialType, CiPipelineSourceTypeOption, Githost } from './types'
 import InfoColourBar from '../common/infocolourBar/InfoColourbar'
 import { ReactComponent as InfoIcon } from '../../assets/icons/info-filled.svg'
+import { reactSelectStyles } from '../CIPipelineN/ciPipeline.utils'
 
 export interface SourceMaterialsProps {
     materials: MaterialType[]
